ENH: Addition of a .travis.yml file
To support auto-compilation on each push and pull request on that project, we add the support of https://travis-ci.org that is linked to github.com
The version of cmake that is available on Travis Cl is 2.8.7. We need at least the version 2.8.9. We build and install cmake 2.8.12 to be able to compile the project