ENH: Update FiberViewerLight from revision 78 to 79
------------------------------------------------------------------------
r79 | fbudin | 2014-10-06 16:21:17 -0400 (Mon, 06 Oct 2014) | 5 lines

ENH: Fiberviewer was not compiling anymore with r23735 (previously?).

There were issues with using the system default CMAKE_C_COMPILER and the CMAKE_CXX_COMPILER to build the extension when Slicer was built with other compilers.

------------------------------------------------------------------------
------------------------------------------------------------------------
r78 | fbudin | 2014-07-30 14:31:22 -0400 (Wed, 30 Jul 2014) | 3 lines

ENH: Specify which version of Qt we are looking for to avoid problems when finding Qt3 and Qt4

------------------------------------------------------------------------